Image-Detection-Bypass-Utility-main/requirements.txt
scikit-image
pyqt5
pillow
numpy
matplotlib
scikit-image
piexif
opencv-python
--extra-index-url https://download.pytorch.org/whl/cu126
#torch==2.8.0+cu126
#torchvision==0.23.0+cu126
scipy
lpips
